feat: refactor into SubjectGrid
